refactor: make `environment` module more readable
